# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EVA — Verified Authorship API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  author: string;
  title: string;
  body: string;
  hash: string;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  walletAddress?: string | null;
  date: string;
}

export interface PostInput {
  author: string;
  title: string;
  body: string;
  hash: string;
  /** @nullable */
  signature?: string | null;
  /** @nullable */
  walletAddress?: string | null;
}

export interface VerifyInput {
  hash: string;
}

export interface VerifyResponse {
  found: boolean;
  post?: Post | null;
}

export interface FeedResponse {
  posts: Post[];
  /** @nullable */
  error?: string | null;
}

export interface PublishResponse {
  success: boolean;
  blobName?: string;
  post?: Post;
}

export interface ErrorResponse {
  error: string;
}