# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EVA — Verified Authorship API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: EVA post operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get published posts
      description: Returns all published posts from Shelby storage (or in-memory store if Shelby is not configured)
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
  /publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish a post
      description: Store a post with its authorship proof to Shelby storage (or in-memory store if Shelby is not configured)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "200":
          description: Post published successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verify:
    post:
      operationId: verifyPost
      tags: [posts]
      summary: Verify a post hash
      description: Look up a SHA-256 hash and return the matching post if found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Post:
      type: object
      properties:
        author:
          type: string
        title:
          type: string
        body:
          type: string
        hash:
          type: string
        signature:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        date:
          type: string
      required:
        - author
        - title
        - body
        - hash
        - date
    PostInput:
      type: object
      properties:
        author:
          type: string
        title:
          type: string
        body:
          type: string
        hash:
          type: string
        signature:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
      required:
        - author
        - title
        - body
        - hash
    VerifyInput:
      type: object
      properties:
        hash:
          type: string
      required:
        - hash
    VerifyResponse:
      type: object
      properties:
        found:
          type: boolean
        post:
          oneOf:
            - $ref: "#/components/schemas/Post"
            - type: "null"
      required:
        - found
    FeedResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        error:
          type: ["string", "null"]
      required:
        - posts
    PublishResponse:
      type: object
      properties:
        success:
          type: boolean
        blobName:
          type: string
        post:
          $ref: "#/components/schemas/Post"
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error