# lib/api-zod/src/generated/types/verifyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EVA — Verified Authorship API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface VerifyResponse {
  found: boolean;
  post?: Post | null;
}